refactor: pull checks into new module
